--- src/pages/Login/Login.tsx
@@ -1,6 +1,0 @@
-import SignIn from '../../components/Popups/WithForm/SignIn/SignIn';
-
-const Login = () => {
-    return <SignIn />;
-};
-export default Login;
