+++ src/pages/SignIn/SignIn.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables.scss';
+@import '../../styles/variables.scss';
 
 .link_registration {
     color: $orange;
@@ -22,7 +22,7 @@
     &__warning {
         min-width: 25px;
         min-height: 25px;
-        background-image: url(../../../../vendor/images/warning.png);
+        background-image: url(../../vendor/images/warning.png);
         background-size: cover;
         background-repeat: no-repeat;
     }
